requirements.txt
pytz
PyYAML
google-cloud-logging
google-cloud-bigquery
requests
boto3
botocore